fix: ensure rendered layers reflect structure and order of Map children
The problem this commit tries to solve can be explained with the
following example.

```
<Map>
  {someSideEffect && (<LayerA/>)}
  <LayerB/>
  <LayerC/>
</Map>
```

`someSideEffect` is initially false, but becomes true after some time.

Initially, the map's layer list contains B and C. When `someSideEffect`
becomes true, it is *appended* to the map's layer list, and so it will
be rendered on top of B and C.

This commit ensures that the map's layers list always reflects the order
of the Map component's children.

It does this by keeping a per-render list of the children in context
(the `registerdLayers` list in context), and checking that it exactly
matches the map's layers list. If it doesn't (i.e. if a layer is added
or re-created), the map's layers list is reset.