feat: re-use canvas between renders
This change replicates the behaviour we had prior to f2030d2, before I
removed it!

It fixes the glitchy zooming and panning issues we've been noticing in
the US election pages recently.

Without this fix, we're creating a new canvas element whenever anything
in the map changes, including zoom level or pan position. Creating a new
canvas for every frame means we're destroying and re-attaching the
mouse/touchmove listeners added by d3-zoom to control zooming and
panning.

Using a single canvas element, as we used to, solves this problem.